Fix runContrastCodeSec_OnSource.sh script. Add Snyk version to runSnyk scripts. Delete all the .xml metadata files for each test case .java file as they aren't needed. All that info is in the expectedresults.csv file for the entire test suite.

+++ scripts/runContrastCodeSec_OnSource.sh
@@ -5,9 +5,17 @@
 # brew tap contrastsecurity/tap
 # brew install contrast
 
+# To scan the source code you have to create a .zip of the src/ directory so src.zip exists in the project's root folder.
+
+if [ -f "./src.zip" ]; then
+
 # Note: you have to do 'contrast auth' first, and successfully authenticate before you can run this.
 benchmark_version=$(scripts/getBenchmarkVersion.sh)
 
-contrast scan -f src --save
+contrast scan -f src.zip --timeout=10000 --save
 mv results.sarif results/Benchmark_$benchmark_version-ContrastCodeSec_OnSrc.sarif
+
+else
+  echo "ERROR: You must create a .zip of the src directory first, before running this script."
+fi
